Fix issues with tslint
