+++ package.json
@@ -15,7 +15,6 @@
     "@types/core-js": "^0.9.35",
     "@types/mocha": "^2.2.34",
     "@types/node": "^6.0.52",
-    "@types/path-to-regexp": "^1.0.29",
     "@types/sinon": "^1.16.33",
     "babel-cli": "^6.18.0",
     "babel-core": "^6.21.0",
@@ -74,7 +73,7 @@
     "most-subject": "^5.3.0",
     "nyc": "^10.0.0",
     "object.entries": "^1.0.4",
-    "path-to-regexp-es6": "0.0.2",
+    "path-to-regexp": "^1.7.0",
     "pre-commit": "^1.2.2",
     "proptypes": "^0.14.3",
     "redux": "^3.6.0",
