+++ src/router/match.ts
@@ -1,7 +1,8 @@
-import Inferno from 'inferno';
-import pathToRegExp from 'path-to-regexp';
 import { decode, flatten, getURLString, isEmpty, mapSearchParams, pathRankSort, toPartialURL } from './utils';
 import { isArray, toArray } from '../shared';
+
+import Inferno from 'inferno';
+import pathToRegExp from 'path-to-regexp';
 
 const cache: Map<string, IMatchRegex> = new Map();
