+++ src/typings/global.d.ts
@@ -1,9 +1,3 @@
-/* tslint:disable */
-declare module 'hoist-non-inferno-statics' {
-	function hoistStatics(connectClass: any, wrappedComponent: any): {[index: string]: any};
-	export = hoistStatics;
-}
-
 declare module 'concat-stream' {
 	function concatStream(func?: any);
 	export default concatStream;
@@ -18,20 +12,30 @@ declare module 'lodash/fp' {
 	export function curry(obj: any): any;
 }
 
+declare module 'lodash/isPlainObject' {
+	function isPlainObject(value: any): boolean;
+	export = isPlainObject;
+}
+
 declare module 'union-type-es' {
 	export default function (obj: any): any;
 }
 
 interface Window {
-	process: any; 
-	__karma__: any; 
+	process: any;
+	__karma__: any;
 	mocha: any;
 }
 
-//noinspection TsLint
+/* tslint:disable */
 declare namespace process {
-	//noinspection TsLint
 	export interface env {
 		NODE_ENV: any;
 	}
+}
+/* tslint:enable */
+
+declare module 'concat-stream-es6' {
+	function concatStream(func?: any);
+	export default concatStream;
 }
